requirements.txt
discord-py-interactions
pyyaml
